+++ database/migrations/2023_12_28_171137_create_voucher_records_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('voucher_records', function (Blueprint $table) {
             $table->id();
-            $table->decimal("cost", 20, 2);
+            $table->integer("cost");
             $table->decimal("quantity", 20, 2);
             $table->foreignId("unit_id");
             $table->foreignId("product_id");
